Rework notifications
- Fix notification serialization
- Switch to a sort of polling by moderators instead of scheduling notifications
- Do not exectute Azure Table batch operation if no item has been added to the batch

+++ TwitchBingoService/Storage/Azure/BingoNotificationEntity.cs
@@ -29,7 +29,7 @@ namespace TwitchBingoService.Storage.Azure
         {
             return new BingoNotification
             {
-                key = Key,
+                key = (ushort)Key,
                 type = (NotificationType)Type,
                 playerId = PlayerId,
             };
@@ -39,9 +39,9 @@ namespace TwitchBingoService.Storage.Azure
 
         public DateTime NotificationTime { get; set; }
 
-        public ushort Key { get; set; }
+        public Int32 Key { get; set; }
 
-        public byte Type { get; set; }
+        public Int32 Type { get; set; }
 
         public string PlayerId { get; set; }
